Breakout CreateCertInfo from Tpm
